Added Managed Identity parameters to namesapce create and update commands
