+++ src/azure-cli/azure/cli/command_modules/eventhubs/custom.py
@@ -11,22 +11,32 @@ from azure.cli.core.profiles import ResourceType
 
 # , resource_type = ResourceType.MGMT_EVENTHUB
 # Namespace Region
-def cli_namespace_create(cmd, client, resource_group_name, namespace_name, location=None, tags=None, sku='Standard', capacity=None, is_auto_inflate_enabled=None, maximum_throughput_units=None, is_kafka_enabled=None, default_action=None):
+def cli_namespace_create(cmd, client, resource_group_name, namespace_name, location=None, tags=None, sku='Standard', capacity=None,
+                         is_auto_inflate_enabled=None, maximum_throughput_units=None, is_kafka_enabled=None,
+                         default_action=None, identity=None, zone_redundant=None, cluster_arm_id=None ):
     # from azure.mgmt.eventhub.models import EHNamespace, Sku
     EHNamespace = cmd.get_models('EHNamespace', resource_type=ResourceType.MGMT_EVENTHUB)
     Sku = cmd.get_models('Sku', resource_type=ResourceType.MGMT_EVENTHUB)
+    Identity = cmd.get_models('Identity', resource_type=ResourceType.MGMT_EVENTHUB)
+    IdentityType = cmd.get_models('IdentityType', resource_type=ResourceType.MGMT_EVENTHUB)
 
-    if cmd.supported_api_version(resource_type=ResourceType.MGMT_EVENTHUB, min_api='2017-04-01'):
+    if cmd.supported_api_version(resource_type=ResourceType.MGMT_EVENTHUB, min_api='2018-01-01-preview'):
+        ehparam=EHNamespace()
+        ehparam.location = location
+        ehparam.tags = tags
+        ehparam.sku = Sku(name=sku, tier=sku, capacity=capacity)
+        ehparam.is_auto_inflate_enabled = is_auto_inflate_enabled
+        ehparam.maximum_throughput_units = maximum_throughput_units
+        ehparam.kafka_enabled = is_kafka_enabled
+        ehparam.zone_redundant = zone_redundant
+        ehparam.cluster_arm_id=cluster_arm_id
+        if identity:
+            ehparam.identity=Identity(type=IdentityType.system_assigned)
+
         client.create_or_update(
             resource_group_name=resource_group_name,
             namespace_name=namespace_name,
-            parameters=EHNamespace(
-                location=location,
-                tags=tags,
-                sku=Sku(name=sku, tier=sku, capacity=capacity),
-                is_auto_inflate_enabled=is_auto_inflate_enabled,
-                maximum_throughput_units=maximum_throughput_units,
-                kafka_enabled=is_kafka_enabled)).result()
+            parameters=ehparam).result()
 
     if default_action:
         netwrokruleset = client.get_network_rule_set(resource_group_name, namespace_name)
@@ -36,9 +46,13 @@ def cli_namespace_create(cmd, client, resource_group_name, namespace_name, locat
     return client.get(resource_group_name, namespace_name)
 
 
-def cli_namespace_update(cmd, client, instance, tags=None, sku=None, capacity=None, is_auto_inflate_enabled=None, maximum_throughput_units=None, is_kafka_enabled=None, default_action=None):
+def cli_namespace_update(cmd, client, instance, tags=None, sku=None, capacity=None, is_auto_inflate_enabled=None,
+                         maximum_throughput_units=None, is_kafka_enabled=None, default_action=None,
+                         user_identity=None, key_source=None, key_properties=None):
+    Encryption = cmd.get_models('Encryption', resource_type=ResourceType.MGMT_EVENTHUB)
+    KeyVaultProperties = cmd.get_models('KeyVaultProperties', resource_type=ResourceType.MGMT_EVENTHUB)
 
-    if cmd.supported_api_version(resource_type=ResourceType.MGMT_EVENTHUB, min_api='2017-04-01'):
+    if cmd.supported_api_version(resource_type=ResourceType.MGMT_EVENTHUB, min_api='2018-01-01-preview'):
         if tags:
             instance.tags = tags
 
@@ -58,6 +72,26 @@ def cli_namespace_update(cmd, client, instance, tags=None, sku=None, capacity=No
         if is_kafka_enabled:
             instance.kafka_enabled = is_kafka_enabled
 
+        if instance.identity and instance.encryption is None:
+            instance.encryption=Encryption()
+
+
+        if key_source:
+            instance.encryption.key_source=key_source
+
+        if key_properties:
+            keyprop = []
+            if len(key_properties) == 2:
+                keyprop.append(KeyVaultProperties(key_name=key_properties[0],key_vault_uri=key_properties[1]))
+                instance.encryption.key_vault_properties = keyprop
+            elif len(key_properties) == 3:
+                keyprop.append(KeyVaultProperties(key_name=key_properties[0],key_vault_uri=key_properties[1],key_version=key_properties[2]))
+                instance.encryption.key_vault_properties = keyprop
+
+        if user_identity:
+            if user_identity == 'None':
+                instance.identity = None
+
         if default_action:
             resourcegroup = instance.id.split('/')[4]
             netwrokruleset = client.get_network_rule_set(resourcegroup, instance.name)
@@ -68,7 +102,7 @@ def cli_namespace_update(cmd, client, instance, tags=None, sku=None, capacity=No
 
 
 def cli_namespace_list(cmd, client, resource_group_name=None):
-    if cmd.supported_api_version(min_api='2017-04-01'):
+    if cmd.supported_api_version(min_api='2018-01-01-preview'):
         if resource_group_name:
             return client.list_by_resource_group(resource_group_name=resource_group_name)
