+++ src/azure-cli/azure/cli/command_modules/eventhubs/tests/latest/test_eventhub_commands_alias_test.py
@@ -1,0 +1,142 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+# AZURE CLI EventHub - NAMESPACE TEST DEFINITIONS
+
+import time
+
+from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer)
+
+
+# pylint: disable=line-too-long
+# pylint: disable=too-many-lines
+
+
+class EHgeorecoveryCURDScenarioTest(ScenarioTest):
+
+    @ResourceGroupPreparer(name_prefix='cli_test_eh_alias')
+    def test_eh_alias(self, resource_group):
+        from azure.mgmt.eventhub.models import ProvisioningStateDR
+        self.kwargs.update({
+            'loc_south': 'SouthCentralUS',
+            'loc_north': 'NorthCentralUS',
+            'rg': resource_group,
+            'namespacenameprimary': self.create_random_name(prefix='eh-nscli', length=20),
+            'namespacenamesecondary': self.create_random_name(prefix='eh-nscli', length=20),
+            'tags': {'tag1: value1', 'tag2: value2'},
+            'sku': 'Standard',
+            'tier': 'Standard',
+            'authoname': self.create_random_name(prefix='cliAutho', length=20),
+            'defaultauthorizationrule': 'RootManageSharedAccessKey',
+            'accessrights': 'Send',
+            'primary': 'PrimaryKey',
+            'secondary': 'SecondaryKey',
+            'aliasname': self.create_random_name(prefix='cliAlias', length=20),
+            'alternatename': self.create_random_name(prefix='cliAlter', length=20),
+            'id': '',
+            'test': ''
+        })
+
+        self.cmd('eventhubs namespace exists --name {namespacenameprimary}',
+                 checks=[self.check('nameAvailable', True)])
+
+        # Create Namespace - Primary
+        self.cmd('eventhubs namespace create --resource-group {rg} --name {namespacenameprimary} --location {loc_south} --tags {tags} --sku {sku}', checks=[self.check('sku.name', self.kwargs['sku'])])
+
+        # Get Created Namespace - Primary
+        self.cmd('eventhubs namespace show --resource-group {rg} --name {namespacenameprimary}', checks=[self.check('sku.name', self.kwargs['sku'])])
+
+        # Create Namespace - Secondary
+        self.cmd('eventhubs namespace create --resource-group {rg} --name {namespacenamesecondary} --location {loc_north} --tags {tags} --sku {sku}', checks=[self.check('sku.name', self.kwargs['sku'])])
+
+        # Get Created Namespace - Secondary
+        getnamespace2result = self.cmd('eventhubs namespace show --resource-group {rg} --name {namespacenamesecondary}', checks=[self.check('sku.name', self.kwargs['sku'])]).get_output_in_json()
+
+        # Create Authoriazation Rule
+        self.cmd('eventhubs namespace authorization-rule create --resource-group {rg} --namespace-name {namespacenameprimary} --name {authoname} --rights {accessrights}', checks=[self.check('name', self.kwargs['authoname'])])
+
+        partnernamespaceid = getnamespace2result['id']
+        self.kwargs.update({'id': partnernamespaceid})
+        # Get Create Authorization Rule
+        self.cmd('eventhubs namespace authorization-rule show --resource-group {rg} --namespace-name {namespacenameprimary} --name {authoname}', checks=[self.check('name', self.kwargs['authoname'])])
+
+        # CheckNameAvailability - Alias
+
+        self.cmd('eventhubs georecovery-alias exists --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}', checks=[self.check('nameAvailable', True)])
+
+        # Create alias
+        self.cmd('eventhubs georecovery-alias set  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname} --partner-namespace {id}')
+
+        # get alias - Primary
+        self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}')
+
+        # get alias - Secondary
+        self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenamesecondary} --alias {aliasname}')
+
+        getaliasprimarynamespace = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # Get Authorization Rule
+        self.cmd('eventhubs georecovery-alias authorization-rule show --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname} --name {authoname}', checks=[self.check('name', self.kwargs['authoname'])])
+
+        # Get Default Authorization Rule
+        self.cmd('eventhubs georecovery-alias authorization-rule list --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}')
+
+        # check for the Alias Provisioning succeeded
+        while getaliasprimarynamespace['provisioningState'] != ProvisioningStateDR.succeeded.value:
+            time.sleep(30)
+            getaliasprimarynamespace = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        while getaliasprimarynamespace['pendingReplicationOperationsCount'] != 0 and getaliasprimarynamespace['pendingReplicationOperationsCount'] is not None:
+            time.sleep(30)
+            getaliasprimarynamespace = self.cmd(
+                'eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # Break Pairing
+        self.cmd('eventhubs georecovery-alias break-pair  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}')
+
+        getaliasafterbreak = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # check for the Alias Provisioning succeeded
+        while getaliasafterbreak['provisioningState'] != ProvisioningStateDR.succeeded.value:
+            time.sleep(30)
+            getaliasafterbreak = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # Create alias
+        self.cmd('eventhubs georecovery-alias set  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname} --partner-namespace {id}')
+
+        getaliasaftercreate = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # check for the Alias Provisioning succeeded
+        while getaliasaftercreate['provisioningState'] != ProvisioningStateDR.succeeded.value:
+            time.sleep(30)
+            getaliasaftercreate = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        while getaliasaftercreate['pendingReplicationOperationsCount'] != 0 and getaliasaftercreate['pendingReplicationOperationsCount'] is not None:
+            time.sleep(30)
+            getaliasaftercreate = self.cmd(
+                'eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenameprimary} --alias {aliasname}').get_output_in_json()
+
+        # FailOver
+        self.cmd('eventhubs georecovery-alias fail-over  --resource-group {rg} --namespace-name {namespacenamesecondary} --alias {aliasname}')
+
+        getaliasafterfail = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenamesecondary} --alias {aliasname}').get_output_in_json()
+
+        # check for the Alias Provisioning succeeded
+        while getaliasafterfail['provisioningState'] != ProvisioningStateDR.succeeded.value:
+            time.sleep(30)
+            getaliasafterfail = self.cmd('eventhubs georecovery-alias show  --resource-group {rg} --namespace-name {namespacenamesecondary} --alias {aliasname}').get_output_in_json()
+
+        # Delete Alias
+        self.cmd('eventhubs georecovery-alias delete  --resource-group {rg} --namespace-name {namespacenamesecondary} --alias {aliasname}')
+
+        time.sleep(30)
+
+        # Delete Namespace - primary
+        self.cmd('eventhubs namespace delete --resource-group {rg} --name {namespacenameprimary}')
+
+        time.sleep(30)
+
+        # Delete Namespace - secondary
+        self.cmd('eventhubs namespace delete --resource-group {rg} --name {namespacenamesecondary}')
