+++ src/azure-cli/setup.py
@@ -84,7 +84,7 @@ DEPENDENCIES = [
     'azure-mgmt-devtestlabs~=4.0',
     'azure-mgmt-dns~=2.1',
     'azure-mgmt-eventgrid~=2.2',
-    'azure-mgmt-eventhub~=3.0.0',
+    'azure-mgmt-eventhub~=4.0.0',
     'azure-mgmt-hdinsight~=1.4.0',
     'azure-mgmt-imagebuilder~=0.4.0',
     'azure-mgmt-iotcentral~=3.0.0',
